third_party/blink/web_tests/http/tests/inspector-protocol/accessibility/tree-updated-events-expected.txt

Test that accessibility tree changes triggers events

Running test: expectEventFiresForAppendedNodes
Nodes updated includes node ID equal to childrenAddRemoveTarget ID? true

Running test: expectEventFiredWhenDataModified
Received update for textnode? true
